requirements.txt
groq
python-dotenv>=0.19.0
streamlit>=1.28.0
langchain>=0.1.0
langchain-groq>=0.0.1
langchain-core>=0.1.0
typing>=3.7.4
textblob>=0.15.3
nltk>=3.8.1
emoji>=2.2.0
plotly>=5.13.0
pandas>=1.5.3
